+++ tests/js/session-tests.js
@@ -34,12 +34,14 @@ function node_require(module) {
 let tmp;
 let fs;
 let execFile;
+let path;
 
 if (isNodeProccess) {
     tmp = node_require('tmp');
     fs = node_require('fs');
     execFile = node_require('child_process').execFile;
     tmp.setGracefulCleanup();
+    path = node_require("path");
 }
 
 
@@ -74,6 +76,14 @@ function promisifiedLogin(server, username, password) {
             }
         });
     });
+}
+
+function copyFileToTempDir(filename) {
+    let tmpDir = tmp.dirSync();
+    let content = fs.readFileSync(filename);
+    let tmpFile = tmp.fileSync({ dir: tmpDir.name });
+    fs.appendFileSync(tmpFile.fd, content);
+    return tmpFile.name;
 }
 
 function runOutOfProcess(nodeJsFilePath) {
@@ -461,6 +471,130 @@ module.exports = {
         });
     },
 
+    testIncompatibleSyncedRealmOpen() {
+        let realm = "sync-v1.realm";
+        if (isNodeProccess) {
+            realm = copyFileToTempDir(path.join(process.cwd(), "data", realm));
+        }
+        else {
+            //copy the bundled RN realm files for the test
+            Realm.copyBundledRealmFiles();
+        }
+
+        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then(user => {
+            return new Promise((resolve, _reject) => {
+                const config = {
+                    path: realm,
+                    sync: {
+                        user,
+                        error : err => cosole.log(err),
+                        url: 'realm://localhost:9080/~/sync-v1'
+                    }
+                };
+
+                Realm.open(config)
+                    .then(realm =>
+                        _reject("Should fail with IncompatibleSyncedRealmError"))
+                    .catch(e => {
+                        if (e.name == "IncompatibleSyncedRealmError") {
+                            const backupRealm = new Realm(e.configuration);
+                            TestCase.assertEqual(backupRealm.objects('Dog').length, 3);
+                            resolve();
+                            return;
+                        }
+
+                        function printObject(o) {
+                            var out = '';
+                            for (var p in o) {
+                              out += p + ': ' + o[p] + '\n';
+                            }
+                            return out;
+                          }
+
+                        _reject("Failed with unexpected error " + printObject(e));
+                    });
+            });
+        });
+    },
+
+    testIncompatibleSyncedRealmOpenAsync() {
+        let realm = "sync-v1.realm";
+        if (isNodeProccess) {
+            realm = copyFileToTempDir(path.join(process.cwd(), "data", realm));
+        }
+        else {
+            //copy the bundled RN realm files for the test
+            Realm.copyBundledRealmFiles();
+        }
+
+        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then(user => {
+            return new Promise((resolve, _reject) => {
+                const config = {
+                    path: realm,
+                    sync: {
+                        user,
+                        error : err => cosole.log(err),
+                        url: 'realm://localhost:9080/~/sync-v1'
+                    }
+                };
+
+                Realm.openAsync(config, (error, realm) => {
+                    if (!error) {
+                        _reject("Should fail with IncompatibleSyncedRealmError");
+                        return;
+                    }
+
+                    if (error.name == "IncompatibleSyncedRealmError") {
+                        const backupRealm = new Realm(error.configuration);
+                        TestCase.assertEqual(backupRealm.objects('Dog').length, 3);
+                        resolve();
+                        return;
+                    }
+
+                    _reject("Failed with unexpected error" + JSON.stringify(error));
+                });
+            });
+        });
+    },
+
+    testIncompatibleSyncedRealmConsructor() {
+        let realm = "sync-v1.realm";
+        if (isNodeProccess) {
+            realm = copyFileToTempDir(path.join(process.cwd(), "data", realm));
+        }
+        else {
+            //copy the bundled RN realm files for the test
+            Realm.copyBundledRealmFiles();
+        }
+
+        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then(user => {
+            return new Promise((resolve, _reject) => {
+                    const config = {
+                        path: realm,
+                        sync: {
+                            user,
+                            error : err => cosole.log(err),
+                            url: 'realm://localhost:9080/~/sync-v1'
+                        }
+                    };
+
+                    try {
+                        const realm = new Realm(config);
+                        _reject("Should fail with IncompatibleSyncedRealmError");
+                    }
+                    catch (e) {
+                        if (e.name == "IncompatibleSyncedRealmError") {
+                            const backupRealm = new Realm(e.configuration);
+                            TestCase.assertEqual(backupRealm.objects('Dog').length, 3);
+                            resolve();
+                            return;
+                        }
+
+                        _reject("Failed with unexpected error" + JSON.stringify(e));
+                    }
+            });
+        });
+    },
 
     testProgressNotificationsForRealmConstructor() {
         if (!isNodeProccess) {
@@ -646,4 +780,111 @@ module.exports = {
                 });
             });
     },
+
+    testPartialSync() {
+        // FIXME: try to enable for React Native
+        if (!isNodeProccess) {
+            return Promise.resolve();
+        }
+
+        const username = uuid();
+        const realmName = uuid();
+
+        return runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
+            .then(() => {
+                Realm.Sync.User.login('http://localhost:9080', username, 'password').then(user1 => {
+                    TestCase.assertDefined(user1, 'user1');
+                    let config1 = {
+                        sync: {
+                            user: user1,
+                            url: `realm://localhost:9080/~/${realmName}`,
+                        },
+                        schema: [{ name: 'Integer', properties: { value: 'int' } }],
+                    };
+                    return new Promise((resolve, reject) => {
+                        return Realm.open(config1)
+                            .then(realm1 => {
+                                for(let i = 0; i < 10; i++) {
+                                    realm1.write(() => {
+                                        realm1.create('Integer', {value: i});
+                                    });
+                                }
+
+                                const progressCallback = (transferred, total) => {
+                                    if (transferred === total) {
+                                        resolve();
+                                    }
+                                }
+                                realm.syncSession.addProgressNotification('upload', 'reportIndefinitely', progressCallback);
+                            })
+                            .then(() => {
+                                realm.close();
+                                realm.deleteFile(config1);
+                                user1.logout();
+                            });
+                    });
+                })
+            })
+            .then(() => {
+                Realm.Sync.User.login('http://localhost:9080', username, 'password').then(user2 => {
+                    TestCase.assertDefined(user2, 'user2');
+                    return new Promise((resolve, reject) => {
+                        let config2 = {
+                            sync: {
+                                user: user2,
+                                url: `realm://localhost:9080/~/${realmName}`,
+                                partial: true,
+                            },
+                            schema: [{ name: 'Integer', properties: { value: 'int' } }],
+                        };
+
+                        const realm2 = new Realm(config2);
+                        realm2.subscribeToObjects('Integer', 'value > 5').then((results, error) => {
+                            return results;
+                        }).then((results) => {
+                            TestCase.assertEqual(results.length, 4);
+                            for(obj in results) {
+                                TestCase.assertTrue(obj.value > 5, '<= 5');
+                            }
+                            resolve();
+                        });
+                        reject();
+                    })
+                })
+            })
+    },
+
+    testClientReset() {
+        // FIXME: try to enable for React Native
+        if (!isNodeProccess) {
+            return Promise.resolve();
+        }
+
+        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then(user => {
+            return new Promise((resolve, _reject) => {
+                var realm;
+                const config = { sync: { user, url: 'realm://localhost:9080/~/myrealm' } };
+                config.sync.error = (sender, error) => {
+                    try {
+                        TestCase.assertEqual(error.code, 7); // 7 -> client reset
+                        TestCase.assertDefined(error.config);
+                        TestCase.assertNotEqual(error.config.path, '');
+                        const original_path = realm.path;
+                        realm.close();
+                        Realm.Sync.initiateClientReset(original_path);
+                        // copy required objects from Realm at error.config.path
+                        resolve();
+                    }
+                    catch (e) {
+                        _reject(e);
+                    }
+                };
+                realm = new Realm(config);
+                const session = realm.syncSession;
+
+                TestCase.assertEqual(session.config.error, config.sync.error);
+                session._simulateError(211, 'ClientReset'); // 211 -> divering histories
+            });
+        });
+    }
 }