+++ CHANGELOG.md
@@ -4,16 +4,20 @@ X.Y.Z Release notes
 * Removed `setAccessToken()`; use `setFeatureToken()` instead.
 
 ### Enhancements
-* None.
+* Support migration from Realms sync 1.0 to sync 2.0 versions
+* Handling of the situation when the client has to reset the Realm due to diverging histories (#795).
+* Added `Realm.subscribeToObjects()` to listen for changes in partially synced Realms.
 
 ### Bug fixes
 * None.
 
-2.0.0 Release notes (2017-9-29)
+
+2.0.0-rc16 Release notes (2017-9-29)
 =============================================================
 ### Breaking changes
 * Upgtading to Realm Core 4.0.1 (bug fixes)
 * Upgrading to Realm Sync 2.0.0-rc26 (sync protocol 22 + bug fixes)
+
 
 2.0.0-rc14 Release notes (2017-9-29)
 =============================================================
@@ -21,7 +25,7 @@ X.Y.Z Release notes
 * Upgrading to Realm Core 4.0.0 and Realm Sync 2.0.0-rc25.
 
 ### Enhancements
-* None.
+* None
 
 ### Bug fixes
 * Configuration of sync file system is not done on module import but later when actually needed by sync (#1351)